reorganización estilos
