+++ frontend-app/src/components/FeedPosts.js
@@ -3,11 +3,11 @@ import React, {useEffect, useState, useContext} from 'react';
 import {AuthContext} from "../utils/AuthFront/context";
 import {CopyToClipboard} from "react-copy-to-clipboard/lib/Component";
 import moment from "moment";
+import {useStyles} from '../styles/styles';
 
 /* COMPONENTS & STYLES */
-import '../App.css';
+import '../styles/App.css';
 import Grid from "@material-ui/core/Grid";
-import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Avatar from '@material-ui/core/Avatar';
 import Divider from '@material-ui/core/Divider';
@@ -18,71 +18,6 @@ import ReplyIcon from '@material-ui/icons/Reply';
 import Modal from '@material-ui/core/Modal';
 
 
-const useStyles = makeStyles(theme =>({
-    postslist:{
-        marginTop: 30,
-        marginLeft: 30,
-        width: '100%',
-    },
-    singlepost:{
-        marginBottom: 20,
-        padding: 10,
-    },
-    profileicon:{
-        width: theme.spacing(6),
-        height: theme.spacing(6),
-        display: 'inline-flex',
-        position:'relative',
-    },
-    authorbox:{
-        display:'flex',
-        position:'relative',
-    },
-    authorinfo:{
-        display:'inline',
-        position:'relative',
-        marginLeft: '10px',
-        marginTop: '-5px',
-    },
-    text:{
-        fontSize: '80%',
-        lineHeight:'2px',
-    },
-    title:{
-        lineHeight:'2px',
-    },
-    postbuttons:{
-        paddingTop:'1px',
-        paddingBottom:'1px',
-    },
-    iconbuttons:{
-        marginRight:'4px',
-    },
-    paper: {
-        position: 'absolute',
-        width: 400,
-        backgroundColor: theme.palette.background.paper,
-        border: '0px solid #000',
-        boxShadow: theme.shadows[5],
-        top: "50%",
-        left: "50%",
-        marginLeft: "-200px",
-        marginTop: "-150px"
-    },
-    popupHeader: {
-        padding: theme.spacing(2, 4, 3),
-        backgroundColor: "#3f51b5",
-        color: "white"
-    },
-    textPadding: {
-        padding: theme.spacing(2, 4, 3),
-        justifyContent: 'center',
-    },
-    loadmore:{
-        position: 'absolute',
-        right: '48%',
-    }
-}));
 
 
 function FeedPosts() {
@@ -93,7 +28,7 @@ function FeedPosts() {
     const {dispatch} = useContext(AuthContext);  // no incluyo state porque no lo estamos usando. reañadir si hiciera falta
 
     const [postsData, setPostsData] = useState([]);
-    const [error, setError] = useState(false);
+    const [, setError] = useState(false);
     const [length, setLength] = useState(5);
 
 
@@ -143,11 +78,6 @@ function FeedPosts() {
                 <Grid item xs={11}>
                     <Grid item xd={10} className={classes.postslist}>
                         {postsData.data && postsData.data.map((data) => {
-                            const style = {
-                                avatar: {
-                                    backgroundColor: data.color
-                                }
-                            };
                             return (
                                 <Paper className={classes.singlepost}>
                                     <Grid item xd={10} className={classes.authorbox}>
