+++ frontend-app/src/components/Homebutton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../App.css';
+import '../styles/App.css';
 import HomeOutlinedIcon from '@material-ui/icons/HomeOutlined';
