+++ frontend-app/src/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import './styles/styles';
 import './index.css';
 import * as serviceWorker from './serviceWorker';
 import App from "./App";
