+++ frontend-app/src/views/FeedPage.js
@@ -1,42 +1,16 @@
 /* BASIC STUFF */
 import React from 'react';
 import {AuthContext} from "../utils/AuthFront/context";
+import {useStyles} from '../styles/styles';
 
 /* COMPONENTS & STYLES */
-import { makeStyles } from '@material-ui/core/styles';
-import '../App.css';
+import '../styles/App.css';
 import Grid from "@material-ui/core/Grid";
 import {Container} from "@material-ui/core";
 import CreatePost from "../components/CreatePost";
 import ProfileInfoFeed from "../components/ProfileInfoFeed";
 import FeedPosts from "../components/FeedPosts";
 
-const useStyles = makeStyles(theme =>({
-    root: {
-        background: 'rgb(241,238,238)',
-        display: 'inline-block',
-        height: 'auto',
-        minHeight: '1500px',
-        width:'100%',
-        paddingTop: '30px',
-        paddingLeft: 200 ,
-        paddingRight: 200 ,
-    },
-    columnCenter:{
-        float: 'left',
-        width: '50%',
-        backgroundColor: 'rgb(241,238,238)',
-        height: 'auto',
-    },
-    columnSides: {
-        float: 'left',
-        width: '25%',
-        backgroundColor: 'rgb(241,238,238)',
-        height: '60em',
-        justifyContent: 'center',
-    },
-
-}));
 
 function FeedPage(){
 
@@ -44,11 +18,11 @@ function FeedPage(){
 
     return (<AuthContext.Consumer>
         {props =>
-            <Container className={classes.root} maxWidth={'xl'}>
-                <div className={classes.columnSides}>
+            <Container className={classes.rootFeed} maxWidth={'xl'}>
+                <div className={classes.columnFeedSides}>
                     <ProfileInfoFeed/>
                 </div>
-                <div className={classes.columnCenter}>
+                <div className={classes.columnFeedCenter}>
                     <Grid container spacing={6}>
                     <CreatePost/>
                     </Grid>
@@ -57,7 +31,7 @@ function FeedPage(){
                     </Grid>
                 </div>
 
-            <div className={classes.columnSides}>
+            <div className={classes.columnFeedSides}>
             </div>
             </Container>
         }
