+++ frontend-app/src/views/Login-Form.js
@@ -12,7 +12,7 @@ import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import {Typography} from '@material-ui/core';
 import TextField from '@material-ui/core/TextField';
-import '../App.css';
+import '../styles/App.css';
 
 function LoginForm({history}) {
